Inlined a class which is only used in one place now
